AI_model/backend/requirements.txt
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
supabase==1.0.3
python-dotenv==1.0.0
fastapi==0.95.2
uvicorn==0.23.1
pydantic==1.10.9
httpx==0.23.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6 